feat: Move AttachmentList helpers into attachment utils
