+++ attachments/utils.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 from main.utils import renderable, is_secretary
 from proposals.models import Proposal
 from studies.models import Study
-from reviews.templatetags.documents_list import DocList
+from reviews.templatetags.documents_list import Container, DocItem
 
 
 class desiredness:
@@ -123,6 +123,99 @@ class AttachmentSlot(renderable):
                 "other_pk": self.attached_object.pk,
             },
         )
+
+
+class AttachmentItem(DocItem):
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # self.attachment = self.slot.attachment.get_correct_submodel()
+
+    @property
+    def comparable(self):
+        if not self.attachment.parent:
+            return False
+        ancestor_proposal = self.slot.get_proposal().parent
+        if not ancestor_proposal:
+            return False
+        direct_ancestors = (
+            [ancestor_proposal] +
+            list(ancestor_proposal.study_set.all())
+        )
+        parent_attachment = self.attachment.parent.get_correct_submodel()
+        if set(parent_attachment.attached_to.all()) & set(direct_ancestors):
+            return True
+
+    @property
+    def compare_url(self,):
+        return reverse(
+            "proposals:compare_attachments",
+            kwargs={
+                "proposal_pk": self.slot.get_proposal().pk,
+                "old_pk": self.attachment.parent.pk,
+                "new_pk": self.attachment.pk,
+            }
+        )
+
+    @property
+    def attachment(self,):
+        return self.slot.attachment.get_correct_submodel()
+
+    def get_link_url(self,):
+        return self.slot.attachment.get_download_url(
+            self.slot.get_proposal(),
+        )
+
+    def get_filename(self,):
+        return self.attachment.upload.original_filename
+
+
+class DocList(list):
+    """
+    A list-like of attachments that can output its contents as containers
+    separated per attached objects.
+    """
+
+    def as_containers(self):
+        containers = []
+        per_item = self.per_item()
+        for item in per_item.keys():
+            if type(item) is Proposal:
+                container = self.make_proposal_container(item)
+            else:
+                # This must be a Study object
+                container = Container(_("Traject {}").format(item.order))
+                container.order = 200 + item.order
+            container.items += [
+                self.make_docitem(slot) for slot in per_item[item]
+            ]
+            containers.append(container)
+        return sorted(containers, key=lambda c: c.order)
+
+    def per_item(self):
+        items = set([slot.attached_object for slot in self])
+        item_dict = {
+            item: [] for item in items
+        }
+        for slot in self:
+            item_dict[slot.attached_object].append(slot)
+        return item_dict
+
+    def make_proposal_container(self, proposal):
+        container = Container(_("Aanvraag"))
+        container.order = 100
+        # The proposal PDF isn't an attachment, so we add it manually
+        proposal_pdf = DocItem(_("Aanvraag in PDF-vorm"))
+        proposal_pdf.link_url = reverse("proposals:pdf", args=(proposal.pk,))
+        container.items.append(proposal_pdf)
+        return container
+
+    def make_docitem(self, slot):
+        docitem = AttachmentItem(
+            slot.kind.name,
+            slot=slot,
+        )
+        return docitem
 
 
 class AttachmentsList(renderable):
